coud_function/requirements.txt
google-cloud-storage==1.27.0
Pillow==7.2.0
